Multi-repository documentation build (#33)
Introduce a new multi-repository GitHub actions workflow that builds the documentation into the 'incontext-site' repository. It's limited to running on `main` commits, and is serialised using the https://github.com/softprops/turnstyle action. Long term, this should probably be an active pull into the site repository instead of a push.

This change also introduces some small documentation additions to test the workflow.