Ensure that TIFF snapshots are converted to JEPGs (#93)
Unfortunately Photos.app still seems to have a bug whereby the Title EXIF key isn't set when exporting photos as JEPGs. It's still present for TIFF exports meaning this is currently the only way to get all the metadata out of Apple Photos. Since TIFF can be pretty unwieldy, this change introduces automatic conversion of TIFF shapshots to JPEG.

It seems that some EXIF data is lost in the process of changing format using ImageMagick's `convert` command, so this change ensures it's copied over from the original file using `exiftool`.