Centralise the EXIF and path parsing to ensure consistency across media types (#95)
This change is motivated by a desire to add better title, description, and location parsing to video file import. It introduces a lightweight schema-based processor which can be used to concisely express EXIF property selection, merges the image and video tags into a common schema, and [currenctly] exposes this functionality through the gallery module.

It includes a drive-by update to the copyright notices and some organisational cleanup in gallery.py.