Support EXIF dates within the new schema module and use this when selecting an appropriate date (#100)
Support EXIF dates within the new schema module and use this when selecting an appropriate date
This change consolidates all the date field selection and parsing into the `METADATA_SCHEMA`. In order to do this, it
moves the ExifTool formatted date parsing into a new schema transform, `EXIFDate`. To ensure that date selection is
consistent across InContext, it also adds some additional date fields to the metadata schema that were previously being
used exclusively for the `manage` command.

This change also adds test data and smoke tests for EXIF handling.