Use ImageMagick for more reliable HEIC image conversion and resizing (#102)
In order to address an issue when resizing HEIC files (the stride of cropped HEIC images was incorrect), this change reverts to using ImageMagick as the primary path for converting and resizing images. In order to do this, the Docker image has been updated to use a source build of ImageMagick within HEIC support.

This change also makes some tentative steps towards a more declarative approach for specifying image thumbnail generation (see the most recent notes in (#10). It intoduces a new class (Glob) which can be used used to concisely represent and perform a filename glob match which can hopefully be used as the basis for a configuration-based modelling of transform operations.