feat: Preview changes using a local server with the `serve` command (#92)
This PR introduces a new command for running a development server that will optionally watch the source directory and trigger rebuilds.

In order to support exposing the server port, it introduces the concept of a preflight plugin that can modify the Docker container options. The change includes a drive-by fix to ensure that the guest is not 'over mounted' leading to IO performance issues, especially on macOS, as it's now possible to determine the site root outside the Docker container.