ci: Prune Docker builds to ensure the cache isn't full (#166)
It's unfortunate that we have to do this, but it seems to be the only way to ensure the build machines don't get completely swamped with images. Perhaps a more elegant approach would be to have a dedicated maintenence task that kept the build machines clean, but it amounts to the same thing and this is at least clear and explicit.